promethea-engine copy 2/pyproject.toml
[tool.poetry]
name = "promethea-engine"
version = "0.1.0"
description = "An autonomous AI trading agent based on a generative world model."
authors = ["Your Name <you@example.com>"]
readme = "ROADMAP.md"
packages = [{include = "trading_agent"}, {include = "promethea"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
pydantic-settings = "*"
sqlalchemy = "*"
psycopg2-binary = "*"
python-dotenv = "*"
pandas = ">=2.1.4,<2.2.0"
polars = {extras = ["pyarrow"], version = "*"}
joblib = "*"
scikit-learn = "*"
tensorflow = "2.16.1"
tensorflow-metal = "1.2.0"
numpy = "1.26.4"
fredapi = "*"
torch = "*"
alpaca-py = "*"
feedparser = "*"
alembic = "*"
pytrends = "*"
google-generativeai = "^0.8.5"
vadersentiment = "^3.3.2"
beautifulsoup4 = "^4.14.2"
aiohttp = "^3.12.15"
requests = "^2.32.5"
curl-cffi = "^0.13.0"
yfinance = "^0.2.66"
praw = "^7.8.1"
google-cloud-pubsub = "^2.22.0"
google-auth = "^2.30.0"
web3 = "^6.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.2,<9.0.0"
pytest-mock = ">=3.15.1,<4.0.0"
polars = ">=1.34.0,<2.0.0"
pytest-anyio = ">=0.0.0-dev"
trio = ">=0.28.0,<1.0.0"
anyio = "^4.11.0"
stable-baselines3 = "^2.7.0"
tomlkit = "^0.13.0"
gymnasium = "^1.2.1"
